requirement.txt
streamlit==1.22.0
pandas==1.5.3
numpy==1.24.2
python-dotenv==1.0.0
plotly==5.13.0
ibm-watson-machine-learning==1.0.357